chore (refs T22614): move loadLibs items to lib/core.
