+++ client/js/directives/saveAndReturn.js
@@ -7,7 +7,7 @@
  * All rights reserved
  */
 
-import { addFormHiddenField } from '@DpJs/lib/FormActions'
+import { addFormHiddenField } from '@DpJs/lib/core/FormActions'
 
 const saveAndReturn = {
   inserted: function (el) {
