+++ client/js/lib/core/CharCount.js
