+++ client/js/lib/core/CheckableItem.js
