+++ client/js/lib/core/Pager.js
