+++ client/js/lib/core/Sticky.js
