+++ demosplan/DemosPlanCoreBundle/Resources/client/js/InitVue.js
@@ -14,7 +14,7 @@ import initGlobalEventListener from '@DpJs/lib/GlobalEventListener'
 import { initStore } from '@DpJs/store/core/initStore'
 import initToggleSideMenu from '@DpJs/lib/ToggleSideMenu'
 import initUserbox from '@DpJs/lib/touchFriendlyUserbox'
-import { loadLibs } from '@DpJs/lib/loadLibs'
+import { loadLibs } from '@DpJs/lib/core/loadLibs'
 import NotificationStoreAdapter from '@DpJs/lib/NotificationStoreAdapter'
 
 function initialize (components = {}, storeModules = {}, apiStoreModules = [], presetStoreModules = {}) {
